:recycle: #187 - refactor: create separate files for loaders and actions
