+++ frontend/src/pages/destructionlist/create/index.ts
@@ -1,1 +1,3 @@
 export * from "./DestructionListCreate";
+export * from "./DestructionListCreate.action";
+export * from "./DestructionListCreate.loader";
