+++ frontend/src/pages/destructionlist/detail/index.ts
@@ -1,1 +1,3 @@
 export * from "./DestructionListDetail";
+export * from "./DestructionListDetail.action";
+export * from "./DestructionListDetail.loader";
