+++ frontend/src/pages/destructionlist/review/index.ts
@@ -1,1 +1,3 @@
 export * from "./DestructionListReview";
+export * from "./DestructionListReview.action";
+export * from "./DestructionListReview.loader";
