+++ frontend/src/pages/landing/index.ts
@@ -1,1 +1,2 @@
 export * from "./Landing";
+export * from "./Landing.loader";
