+++ frontend/src/pages/login/index.ts
@@ -1,1 +1,2 @@
 export * from "./Login";
+export * from "./Login.action";
