+++ frontend/src/pages/logout/index.ts
@@ -1,1 +1,1 @@
-export * from "./Logout";
+export * from "./Logout.loader";
